### 2023-04-29 #5425 @editorialbot
[PRE REVIEW]: Tashaphyne: A Python package for Arabic Light Stemming
**Submitting author:** @linuxscout (<a href="http://orcid.org/0000-0001-9960-5533">Taha Zerrouki</a>)
**Repository:** https://github.com/linuxscout/tashaphyne/
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.3.6
