### 2023-04-29 #5425 @arfon
:wave: @linuxscout, many thanks for this submission to JOSS. I'm trying to understand the software effort behind this work, but it looks like many of the files include a large number of variables to help with parsing Arabic characters e.g., [like this](https://github.com/linuxscout/tashaphyne/blob/master/tashaphyne/verb_stamp_const.py#L42-L13995).

Could you please link here the main _software_ contributions of this package to help guide the scope review? Also, it looks like there are some files included in the repository that might have been copied in from another project (I think I can see a different creator/license in places). Please also help me understand what's happening here too. Many thanks!
