### 2023-04-29 #5425 @linuxscout
Hi @arfon,

I would like to express my gratitude for your valuable comments regarding our work. Our initial focus was on providing a comprehensive suite of tools and libraries that could efficiently handle Arabic texts and perform various natural language processing tasks. One of the tools that we developed early on was the Tashaphyne stemmer.

As we were working on the stemmer, we realized that certain aspects could be further improved by developing separate packages, such as an Arabic stopwords package. Additionally, PyArabic, which encompasses basic text handling routines, is also an essential package.

While we kept the Tashaphyne stemmer more straightforward and less reliant on other packages, we kept more code for legacy uses because both developers and researchers made extensive use of it. In order to create a high-quality stemmer that many users have adopted, we experimented with various techniques and algorithms.

Our primary contribution is the stemming algorithm contained within stemming.py.

Once again, thank you for your comments.
