+++ docs/paper/paper.bib
@@ -50,7 +50,8 @@ pagetotal = {125}
   year = {2016},
   number	= {6},
   pages = {17--28},
-  volume	= {6}
+  volume	= {6},
+  doi={10.5121/ijdkp.2016.6602}
 }
 
 @PhDThesis{abozinadah2017detecting,
@@ -70,7 +71,8 @@ pagetotal = {125}
    11th International Conference on},
   year = {2014},
   organization	= {IEEE},
-  pages = {148--155}
+  pages = {148--155},
+  doi={10.1109/aiccsa.2014.7073192}
 }
 
 @Article{al2018comprehensive,
@@ -78,7 +80,8 @@ pagetotal = {125}
   author	= {AlAyyoub, Mahmoud and Khamaiseh, Abed Allah and Jararweh,
    Yaser and Al-Kabi, Mohammed N},
   journal	= {Information Processing \& Management},
-  year = {2018}
+  year = {2018},
+doi={10.1016/j.ipm.2018.07.006}
 }
 
 @article{al2019improving,
@@ -96,7 +99,8 @@ pagetotal = {125}
   booktitle={International Conference on Human-Computer Interaction},
   pages={3--14},
   year={2019},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-21902-4_1}
 }
 
 @InProceedings{albukhitan2017arabic,
@@ -106,7 +110,8 @@ pagetotal = {125}
    Intelligence},
   year = {2017},
   organization	= {ACM},
-  pages = {1138--1142}
+  pages = {1138--1142},
+  doi={10.1145/3106426.3109052}
 }
 
 @article{alhaj2019study,
@@ -115,7 +120,8 @@ pagetotal = {125}
   journal={IEEE Access},
   volume={7},
   pages={32664--32671},
-  year={2019}
+  year={2019},
+ doi={10.1109/access.2019.2903331}
 }
 
 @PhDThesis{almuqhim2016,
@@ -155,7 +161,8 @@ pagetotal = {125}
   volume={10},
   number={2},
   pages={256--262},
-  year={2019}
+  year={2019},
+  doi={10.14569/ijacsa.2019.0100234}
 }
 
 @Article{brahimi2016data,
@@ -185,7 +192,8 @@ pagetotal = {125}
   journal	= {International Journal of Computer Applications},
   year = {2015},
   number	= {8},
-  volume	= {125}
+  volume	= {125},
+doi={10.5120/ijca2015906129}
 }
 
 @Article{el2015cbas,
@@ -193,7 +201,8 @@ pagetotal = {125}
   author	= {ElDefrawy, Mahmoud and El-Sonbaty, Yasser and Belal,
    Nahla A},
   journal	= {arXiv preprint arXiv:1611.00027},
-  year = {2015}
+  year = {2015},
+  doi={10.5121/ijnlc.2015.4301}
 }
 
 @InProceedings{el2015enhancing,
@@ -215,7 +224,8 @@ pagetotal = {125}
    and Informatics},
   year = {2016},
   organization	= {Springer},
-  pages = {32--41}
+  pages = {32--41},
+  doi={10.1007/978-3-319-48308-5_4}
 }
 
 @Article{el2016rule,
@@ -226,7 +236,8 @@ pagetotal = {125}
   year = {2016},
   number	= {8},
   pages = {2883--2891},
-  volume	= {41}
+  volume	= {41},
+  doi={10.1007/s13369-016-2029-2}
 }
 
 @InProceedings{el2017efficient,
@@ -236,7 +247,8 @@ pagetotal = {125}
   booktitle	= {Intelligent Systems Conference (IntelliSys), 2017},
   year = {2017},
   organization	= {IEEE},
-  pages = {870--878}
[…]
