+++ docs/paper/paper.bib
@@ -278,7 +278,7 @@ doi={10.5120/ijca2015906129}
   pages={453--458},
   year={2019},
   organization={IEEE},
-  doi={https://github.com/linuxscout/tashaphyne/blob/master/tashaphyne/stemming.py}
+  doi={10.1109/jeeit.2019.8717387}
 }
 
 @Article{hussein2016arabic,
