+++ docs/paper/paper.bib
@@ -513,3 +513,127 @@ doi={10.1145/3320326.3320399}
   doi={10.32604/cmc.2021.016155}
 }
 
+@article{alqahtani2023comparative,
+  title={A Comparative Study of Effective Domain Adaptation Approaches for Arabic Sentiment Classification},
+  author={Alqahtani, Yathrib and Al-Twairesh, Nora and Alsanad, Ahmed},
+  journal={Applied Sciences},
+  volume={13},
+  number={3},
+  pages={1387},
+  year={2023},
+  publisher={MDPI},
+  doi={10.3390/app13031387}
+}
+@article{AMEUR2023100517,
+title = {ChildProtect: A parental control application for tracking hostile surfing content},
+journal = {Entertainment Computing},
+volume = {44},
+pages = {100517},
+year = {2023},
+issn = {1875-9521},
+doi = {https://doi.org/10.1016/j.entcom.2022.100517},
+url = {https://www.sciencedirect.com/science/article/pii/S1875952122000416},
+author = {Hanen Ameur and Amal Rekik and Salma Jamoussi and Abdelmajid Ben Hamadou}
+}
+
+@article{erramisentiment2023,
+author = {Mouaad, Errami and Ouassil, Mohamed Amine and Rachidi, Rabia and Cherradi, Bouchaib and Hamida, Soufiane and Raihani, Abdelhadi},
+year = {2023},
+month = {04},
+pages = {315-325},
+title = {Sentiment Analysis on Moroccan Dialect based on ML and Social Media Content Detection},
+volume = {14},
+journal = {International Journal of Advanced Computer Science and Applications},
+doi = {10.14569/IJACSA.2023.0140347}
+}
+
+@article{muaad2022arabic,
+  title={Arabic document classification: performance investigation of preprocessing and representation techniques},
+  author={Muaad, Abdullah Y and Davanagere, Hanumanthappa Jayappa and Guru, DS and Benifa, JV Bibal and Chola, Channabasava and AlSalman, Hussain and Gumaei, Abdu H and Al-antari, Mugahed A},
+  journal={Mathematical Problems in Engineering},
+  volume={2022},
+  pages={1--16},
+  year={2022},
+  publisher={Hindawi Limited},
+  doi={10.1155/2022/3720358}
+}
+
+@INPROCEEDINGS{hijazi2022review,
+  author={Hijazi, Musab Mustafa and Zeki, Akram and Ismail, Amelia},
+  booktitle={2022 International Arab Conference on Information Technology (ACIT)}, 
+  title={A Review Study on Arabic Text Classification}, 
+  year={2022},
+  volume={},
+  number={},
+  pages={1-13},
+  doi={10.1109/ACIT57182.2022.9994124}
+  }
+@article{abdulsalam2022,
+    doi = {10.1371/journal.pone.0272269},
+    author = {Abdul Salam, Mustafa AND El-Fatah, Mohamed Abd AND Hassan, Naglaa Fathy},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {Automatic grading for Arabic short answer questions using optimized deep learning model},
+    year = {2022},
+    month = {08},
+    volume = {17},
+    url = {https://doi.org/10.1371/journal.pone.0272269},
+    pages = {1-41},
+    number = {8},
+}
+@article{ALAHMADI20222651,
+title = {TAAM: Topic-aware abstractive arabic text summarisation using deep recurrent neural networks},
+journal = {Journal of King Saud University - Computer and Information Sciences},
+volume = {34},
+number = {6, Part A},
+pages = {2651-2665},
+year = {2022},
[…]
