+++ docs/paper/paper.bib
@@ -463,8 +463,7 @@ doi={10.1145/3320326.3320399}
 	volume       = 8,
 	number       = 84,
 	pages        = 4886,
-	doi          = {10.21105/joss.04886},
-	url          = {https://doi.org/10.21105/joss.04886}
+	doi          = {10.21105/joss.04886}
 }
 
 @InProceedings{taghva2005arabic,
@@ -531,7 +530,7 @@ volume = {44},
 pages = {100517},
 year = {2023},
 issn = {1875-9521},
-doi = {https://doi.org/10.1016/j.entcom.2022.100517},
+doi = {10.1016/j.entcom.2022.100517},
 url = {https://www.sciencedirect.com/science/article/pii/S1875952122000416},
 author = {Hanen Ameur and Amal Rekik and Salma Jamoussi and Abdelmajid Ben Hamadou}
 }
@@ -577,7 +576,6 @@ doi = {10.14569/IJACSA.2023.0140347}
     year = {2022},
     month = {08},
     volume = {17},
-    url = {https://doi.org/10.1371/journal.pone.0272269},
     pages = {1-41},
     number = {8},
 }
@@ -589,7 +587,7 @@ number = {6, Part A},
 pages = {2651-2665},
 year = {2022},
 issn = {1319-1578},
-doi = {https://doi.org/10.1016/j.jksuci.2022.03.026},
+doi = {10.1016/j.jksuci.2022.03.026},
 url = {https://www.sciencedirect.com/science/article/pii/S1319157822001100},
 author = {Dimah Alahmadi and Arwa Wali and Sarah Alzahrani}
 }
