### 2023-06-22 #5425 @samhforbes
Hi @linuxscout thanks for submitting to JOSS, and I'm looking forward to digging into this interesting work. I am currently asking some potential reviewers, but if you have any other reviewer suggestions based on out list of reviewers above, please let me know their handles so I can check them out (without tagging them in please).
