### 2023-08-15 #5425 @drussellmrichie
I'm not quite sure I'm ideal for this: I'm not an Arabic speaker and I'm not sure I have the software development knowledge to evaluate the quality of a Python package. And, the rest of my plate is rather full right now. I think I'm going to have to pass. Sorry. 😦
