### 2023-08-28 #5425 @sdruskat
> 👋 @sdruskat @frankier would any of you be willing to review this submission for JOSS?

👋 @samhforbes and thanks for your patience while I was on leave. The topic of this paper is unfortunately well outside my area of expertise, and so I'll have to pass as well. Sorry, and all the best.
