+++ docs/source/index.rst
@@ -1,0 +1,21 @@
+.. Tashaphyne documentation master file, created by
+   sphinx-quickstart on Thu Jan  5 15:36:05 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to Tashaphyne's documentation!
+=====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+   
+   README 
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
