### 2023-10-04 #5425 @tresoldi
Hi @samhforbes . As much as I would like to help, I must unfortunately pass it: my knowledge of Arabic is close to zero. In extreme, I could potentially accept to evaluate the software itself and the NLP side of it if you need it, provided that somebody else addresses the linguistic matter.
