### 2023-10-18 #5425 @samhforbes
Hi @linuxscout do you have any subject matter experts you might be able to recommend here (intended users for example) especially some who speak Arabic? I have been reaching out to some off thread but have struggled to get responses, especially over summer.
