### 2023-11-14 #5425 @samhforbes
Hi @amitkumarj441, @kikarimullah I know you previously reviewed a related package for us. Would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
