### 2023-11-15 #5425 @kikarimullah
Sure, no problem.

Kamran

Dr Kamran I. Karimullah | Lecturer in Islamic Thought | Department of Religions and Theology | School of Arts, Languages and Cultures | University of Manchester | +44 161 275 0475 | Samuel Alexander Building S1.12 | Office hours: Wednesdays 11-1 or by appointment | https://zoom.us/j/5178765556
Genealogies of Knowledge Research Network https://genealogiesofknowledge.net/research-network/
