### 2023-11-15 #5425 @linuxscout
> Hi @linuxscout do you have any subject matter experts you might be able to recommend here (intended users for example) especially some who speak Arabic? I have been reaching out to some off thread but have struggled to get responses, especially over summer.

Hi Sam,

Okay, no worries.

I'd want to continue with the review.

I understand how difficult it is to find reviewers. Thank you very much.

best wishes
