### 2023-11-16 #6063 @editorialbot
[REVIEW]: Tashaphyne: A Python package for Arabic Light Stemming
**Submitting author:** @linuxscout (<a href="http://orcid.org/0000-0001-9960-5533">Taha Zerrouki</a>)
**Repository:** https://github.com/linuxscout/tashaphyne/
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.7
