### 2024-01-08 linuxscout/tashaphyne#8 @SamHames
Use Python 3 syntax examples in the documentation
Just a minor note that it would be nice if the examples used in the documentation used Python 3 syntax (print as a function, no need for the unicode str syntax u'') to enable copy and paste and less confusion for people who might not have ever used Python 2 before. Python 2 went end-of-life four years ago and Python 3 is now the dominant platform by far.

https://github.com/openjournals/joss-reviews/issues/6063
