### 2024-01-08 linuxscout/tashaphyne#9 @SamHames
Example needed: tokenising and working with text longer than a single word
The stemming and related functions for single words are well documented, but most of the time the input for any kind of text processing is going to be larger units (sentences, paragraphs, whole documents etc). I think there needs to be at least a short example of a recommended way to handle this including tokenisation (ie, usage of the included tokenisation methods?) in the readme to highlight this functionality and use case.

If it's your assumption that most people using this package will be tokenising in a separate process and passing only tokens into tashaphyne then I think it word be worth including that in the documentation as well to make expectations/usage even more clear.

https://github.com/openjournals/joss-reviews/issues/6063
