### 2024-01-08 linuxscout/tashaphyne#10 @SamHames
Community guidelines (JOSS checklist)
The JOSS checklist has this section on community guidelines: _Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_

I think from context the answer to all three of these is "use github", but I think it's worth spelling out what people can do - especially people new to programming may not be familiar with github or issue tracking and a more explicit instruction in the readme like "please report any issues on github at URL" can be very helpful.

As a related note, there's a Feedback link in the `readme.md` table to http://tahadz.com/tashaphyne/contact which returns not found.

https://github.com/openjournals/joss-reviews/issues/6063
