### 2024-01-14 f5ac7c5
Update README.md
Add Example to stem a text
