### 2024-01-14 linuxscout/tashaphyne#9 @linuxscout
Thank you for your remark,
I added an example for [stemming text here](https://github.com/linuxscout/tashaphyne/blob/master/README.md#stemming-a-text).
Thanks.
