+++ README.md
@@ -99,7 +99,7 @@ Stemming function: stem an Arabic word and return a stem. This function stores i
 ```python
 >>> from tashaphyne.stemming import ArabicLightStemmer
 >>> ArListem = ArabicLightStemmer()
->>> word = u'أفتضاربانني'
+>>> word = 'أفتضاربانني'
 >>> # stemming word
 ... stem = ArListem.light_stem(word)
 >>> # extract stem
@@ -133,7 +133,7 @@ Stemming function: stem an Arabic word and return a stem. This function stores i
 أفت-انني
 >>> # get affix tuple
 ... print(ArListem.get_affix_tuple())
-{'prefix': u'أفت', 'root': u'', 'stem': u'', 'suffix': u'أفتضاربانني'}
+{'prefix': 'أفت', 'root': '', 'stem': '', 'suffix': 'أفتضاربانني'}
 >>> # star words
 ... print(ArListem.get_starword())
 أفت*ا**انني
@@ -169,7 +169,7 @@ Generate a list of all possible segmentation positions (left, right) of the trea
 
 ```python
 
->>> word = u'أفتضاربانني'
+>>> word = 'أفتضاربانني'
 
 >>> # Detect all possible segmentation
 ... print(ArListem.segment(word))
@@ -181,12 +181,12 @@ set([(2, 7), (3, 8), (0, 8), (2, 9), (2, 8), (3, 10), (2, 11), (1, 8), (0, 7), (
 
 >>> # get affix list
 ... print(ArListem.get_affix_list())
-[{'prefix': u'أف', 'root': u'ضرب', 'stem': u'تضارب', 'suffix': u'انني'},
- {'prefix': u'أفت', 'root': u'ضرب', 'stem': u'ضاربا', 'suffix': u'نني'},
- {'prefix': u'', 'root': u'أفضرب', 'stem': u'أفتضاربا', 'suffix': u'نني'}, 
- {'prefix': u'أف', 'root': u'ضربن', 'stem': u'تضاربان', 'suffix': u'ني'}, 
- {'prefix': u'أف', 'root': u'ضرب', 'stem': u'تضاربا', 'suffix': u'نني'}, 
- {'prefix': u'أفت', 'root': u'ضربنن', 'stem': u'ضاربانن', 'suffix': u'ي'}, ...]
+[{'prefix': 'أف', 'root': 'ضرب', 'stem': 'تضارب', 'suffix': 'انني'},
+ {'prefix': 'أفت', 'root': 'ضرب', 'stem': 'ضاربا', 'suffix': 'نني'},
+ {'prefix': '', 'root': 'أفضرب', 'stem': 'أفتضاربا', 'suffix': 'نني'}, 
+ {'prefix': 'أف', 'root': 'ضربن', 'stem': 'تضاربان', 'suffix': 'ني'}, 
+ {'prefix': 'أف', 'root': 'ضرب', 'stem': 'تضاربا', 'suffix': 'نني'}, 
+ {'prefix': 'أفت', 'root': 'ضربنن', 'stem': 'ضاربانن', 'suffix': 'ي'}, ...]
 >>> 
 ```
 * segment() / get_segment_list()
@@ -209,8 +209,8 @@ You can modify and customize  the default affixes list by
 ```python
 >>> import tashaphyne.stemming
 
[…]
