### 2024-01-14 a350f49
bring examples to python3 syntax
