+++ docs/requirements.txt
@@ -1,1 +1,1 @@
-
+sphinx_rtd_theme
