+++ docs/source/conf.py
@@ -55,7 +55,8 @@ templates_path = ['_templates']
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+# ~ source_suffix = '.rst'
+source_suffix = '.md'
 
 # The master toctree document.
 master_doc = 'README'
