+++ docs/source/README.md
@@ -1,297 +1,1 @@
-# Tashaphyne
-![downloads](https://img.shields.io/pypi/dm/tashaphyne?style=plastic)
-
-
-**Tashaphyne**: Arabic Light Stemmer تاشفين: التجذيع الخفيف للنصوص العربية
-
-**Tashaphyne** is an Arabic light stemmer and segmentor. It mainly supports light stemming (removing prefixes and suffixes) and gives all possible segmentations. It uses a modified finite state automaton, which allows it to generate all segmentations.
-
-It offers stemming and root extraction at the same time, unlike the Khoja stemmer, ISRI stemmer, Assem stemmer, and Farasa stemmer.
-
-**Tashaphyne** comes with default prefixes and suffixes, and accepts the use of customized prefixes and suffixes lists, which allow it to handle more aspects and make customized stemmers without changing code.
-
-**Tashaphyne** is a python library, it's available as a demo on  [Mishkal](http://tahadz.com/mishkal), choose Tools/Analysis and as source code on [Github](http://github.com/linuxscout/tashaphyne) 
-
-  Developpers:  Taha Zerrouki: http://tahadz.com
-    taha dot zerrouki at gmail dot com
-
-Features |   value
----------|---------------------------------------------------------------------------------
-Authors  | [Authors.md](https://github.com/linuxscout/tashaphyne/master/AUTHORS.md)
-Release  | 0.3.6 
-License  |[GPL](https://github.com/linuxscout/tashaphyne/master/LICENSE)
-Tracker  |[linuxscout/tashaphyne/Issues](https://github.com/linuxscout/tashaphyne/issues)
-Website  |[https://pypi.python.org/pypi/Tashaphyne](https://pypi.python.org/pypi/Tashaphyne)
-Doc      |[package Documentaion](https://tashaphyne.readthedocs.io/)
-Source  |[Github](http://github.com/linuxscout/tashaphyne)
-Download  |[sourceforge](http://tashaphyne.sourceforge.net)
-Feedbacks  |[Comments](http://tahadz.com/contact.html)
-Accounts  |[@Twitter](https://twitter.com/linuxscout)  [@Sourceforge](http://sourceforge.net/projects/tashaphyne/)
-
-
-
-## Citation
-If you would cite it in academic work, can you use this citation
-```
-T. Zerrouki‏, Tashaphyne, Arabic light stemmer‏,  https://pypi.python.org/pypi/Tashaphyne/0.2
-```
-or in bibtex format
-```bibtex
-@misc{zerrouki2012tashaphyne,
-  title={Tashaphyne, Arabic light stemmer},
-  author={Zerrouki, Taha},
-  url={https://pypi.python.org/pypi/Tashaphyne/0.2},
-  year={2012}
-}
-```
-
-
-##   مزايا
- - تجذيع الكلمة العربية إلى أبسط جذع ممكن
- - إمكانية استخراج الجذر
- - تقطيع الكلمة إلى جميع الحالات الممكنة.
- - تنميط الكلمة ( توحيد الحروف ذات الأشكال المختلفة.
- - قائمة مسبقة للزوائد العربية، وحروف الزيادة
- -إمكانية ضبط إعدادات المجذع والمقطع، من خلال تعديل قوائم الزوائد.
- 
-## Features
- - Arabic word Light Stemming.
- - Root Extraction.
- - Word Segmentation 
- - Word normalization
[…]
