+++ README.rst
@@ -30,7 +30,7 @@ dot com
 | Authors | `Authors.md <https://github.com/linuxscout/tashaphyne/master/AUT |
 |         | HORS.md>`__                                                      |
 +---------+------------------------------------------------------------------+
-| Release | 0.3.6                                                            |
+| Release | 0.3.5                                                            |
 +---------+------------------------------------------------------------------+
 | License | `GPL <https://github.com/linuxscout/tashaphyne/master/LICENSE>`_ |
 |         | _                                                                |
@@ -47,7 +47,7 @@ dot com
 | Downloa | `sourceforge <http://tashaphyne.sourceforge.net>`__              |
 | d       |                                                                  |
 +---------+------------------------------------------------------------------+
-| Feedbac | `Comments <http://tahadz.com/tashaphyne/contact>`__              |
+| Feedbac | `Comments <http://tahadz.com/contact.html>`__                    |
 | ks      |                                                                  |
 +---------+------------------------------------------------------------------+
 | Account | [@Twitter](https://twitter.com/linuxscout)                       |
@@ -134,52 +134,57 @@ root.
 
 .. code:: python
 
+    >>> #make propre display for unicode
+    ... import pyarabic.arabrepr
+    >>> arepr = pyarabic.arabrepr.ArabicRepr()
+    >>> repr = arepr.repr
+    >>> 
     >>> from tashaphyne.stemming import ArabicLightStemmer
     >>> ArListem = ArabicLightStemmer()
     >>> word = u'أفتضاربانني'
     >>> # stemming word
     ... stem = ArListem.light_stem(word)
     >>> # extract stem
-    ... print(ArListem.get_stem())
+    ... print ArListem.get_stem()
     ضارب
     >>> # extract root
-    ... print(ArListem.get_root())
+    ... print ArListem.get_root()
     ضرب
     >>> 
     >>> # get prefix position index
-    ... print(ArListem.get_left())
+    ... print ArListem.get_left()
     3
     >>> # get prefix 
-    ... print(ArListem.get_prefix())   
+    ... print ArListem.get_prefix()    
     أفت
     >>> # get prefix with a specific index
-    ... print(ArListem.get_prefix(2))
+    ... print ArListem.get_prefix(2)    
     أف
     >>> 
     >>> # get suffix position index
-    ... print(ArListem.get_right())
+    ... print ArListem.get_right()
     7
     >>> # get suffix 
-    ... print(ArListem.get_suffix())
+    ... print ArListem.get_suffix()    
     انني
     >>> # get suffix with a specific index
-    ... print(ArListem.get_suffix(10))    
+    ... print ArListem.get_suffix(10)    
     ي
     >>> # get affix
-    >>> print(ArListem.get_affix())
+    >>> print ArListem.get_affix()
     أفت-انني
     >>> # get affix tuple
-    ... print(ArListem.get_affix_tuple())    
+    ... print repr(ArListem.get_affix_tuple())    
     {'prefix': u'أفت', 'root': u'', 'stem': u'', 'suffix': u'أفتضاربانني'}
     >>> # star words
[…]
