### 2024-01-14 linuxscout/tashaphyne#8 @linuxscout
Thank you, for your comments, I fixed it, in [ReadTheDocs](https://tashaphyne.readthedocs.io/en/latest/).
