### 2024-01-14 #6063 @linuxscout
> Everything looks pretty well put together to me.
>
> I've opened a few minor issues with ([linuxscout/tashaphyne#9](https://github.com/linuxscout/tashaphyne/issues/9), [linuxscout/tashaphyne#10](https://github.com/linuxscout/tashaphyne/issues/10)) being blockers for ticking off the last two items.
>
> I think [linuxscout/tashaphyne#10](https://github.com/linuxscout/tashaphyne/issues/10) should be taken care of eventually but has a bit more surface area and isn't a blocker for this review.

Thank you for your comments, I fixed the issues.
