### 2024-01-14 #6063 @linuxscout
@kikarimullah
Thank you for your comments,

Tashaphyne is a light stemmer for arabic word, it extract stems, which are different from lemmas.
Some stemming result can be different from lemmas, and can give approximate results due to ambiguity, especially with high inflected language as Arabic.
In other hand, Lemmatization aims to reduce words to their base or dictionary form, called the "lemma." This process involves considering the context of the word and its part of speech to produce a valid word.
 For instance, the lemmatization of "running" would be "run," and the lemmatization of "better" would be "good."

Thanks

>
> I tested the stemming and root on an Arabic sentence and I wanted to get @linuxscout comments on the output:
>
> text = "أفترى على الله كذبا ام به جنة بل الذين لا يؤمنون بالاخرة في العذاب والضلال البعيد"
>
> TOKEN: أفترى STEM: رى ROOT: ترر TOKEN: على STEM: على ROOT: على TOKEN: الله STEM: له ROOT: لهه TOKEN: كذبا STEM: ذبا ROOT: كذب TOKEN: ام STEM: ام ROOT: م TOKEN: به STEM: به ROOT: به TOKEN: جنة STEM: جن ROOT: جنن TOKEN: بل STEM: بل ROOT: بل TOKEN: الذين STEM: الذين ROOT: الذين TOKEN: لا STEM: لا ROOT: لا TOKEN: يؤمنون STEM: ؤمن ROOT: ءمن TOKEN: بالاخرة STEM: اخر ROOT: خرر TOKEN: في STEM: في ROOT: في TOKEN: العذاب STEM: عذاب ROOT: عذب TOKEN: والضلال STEM: ضلال ROOT: ضلل TOKEN: البعيد STEM: بعيد ROOT: بعد
>
> I wanted to ask about whether he's done any measurement of the accuracy of the tashaphyne results compared with the results when the stems and roots are identified by hand.
