### 2024-01-25 #6063 @samhforbes
Minor edits
- line 38: change to: Tashaphyne can be found at PyPi.org index, (remove the period before the comma)
- line 38: change to: it is available as a demo on Mikshal (add a)
