### 2024-01-26 #6063 @linuxscout
Hi @samhforbes,

- [ ] I fixed mentioned typos errors.
**-version 0.3.7 **
 **https://zenodo.org/doi/10.5281/zenodo.10570793**
