+++ README.md
@@ -19,7 +19,7 @@ It offers stemming and root extraction at the same time, unlike the Khoja stemme
 Features |   value
 ---------|---------------------------------------------------------------------------------
 Authors  | [Authors.md](https://github.com/linuxscout/tashaphyne/master/AUTHORS.md)
-Release  | 0.3.6 
+Release  | 0.3.7 
 License  |[GPL](https://github.com/linuxscout/tashaphyne/master/LICENSE)
 Tracker  |[linuxscout/tashaphyne/Issues](https://github.com/linuxscout/tashaphyne/issues)
 Website  |[https://pypi.python.org/pypi/Tashaphyne](https://pypi.python.org/pypi/Tashaphyne)
