### 2024-01-30 #6063 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
